refactor(blackbox): move log picker to its own toolbar panel after workspace
Relocate the multi-log picker out of the header dialog's title bar into a
dedicated panel in the top controls row, right after the Workspace selector,
so it is persistently accessible in every view.
